Re-use `UnpredictImageU8` and implement for 16bpp
We suspect the issue with 16bit tiff files are related to the predictor.
In the case where the loader fails, the tiff file uses the LZW
compression and predictor==2. It seems that in 16bit mode the predictor
still assumes 8bit data, and hence the bug.

As a first step we rewrite the code to use the `UnpredictImageU8`
function in more places, so we can fix the issue in less places.

Then implemented `UnpredictImageU8` for 16bpp images as well.